# Fix bugprone-unused-return-value clang-tidy warning